backend/requirements.txt
tortoise-orm==0.19.3
asyncpg==0.27.0
uvicorn==0.22.0
fastapi==0.103.1
